% Generated by roxygen2 (4.0.1): do not edit by hand
\name{doxygen_run}
\alias{doxygen_run}
\title{Run doxygen}
\usage{
doxygen_run(input, output)
}
\arguments{
\item{input}{Vector of directories to search for inut files}

\item{output}{Output directory.  xml will be in
\code{file.path(output, "xml")}.}
}
\description{
Run doxygen
}
\author{
Rich FitzJohn
}

